MAINT: Refactor casts.c to use cpp templates
