+++ numpy/_core/meson.build
@@ -1134,7 +1134,7 @@ src_multiarray = multiarray_gen_headers + [
   'src/multiarray/scalarapi.c',
   'src/multiarray/shape.c',
   'src/multiarray/strfuncs.c',
-  'src/multiarray/stringdtype/casts.c',
+  'src/multiarray/stringdtype/casts.cpp',
   'src/multiarray/stringdtype/dtype.c',
   'src/multiarray/stringdtype/utf8_utils.c',
   'src/multiarray/stringdtype/static_string.c',
